% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{auction_results}
\alias{auction_results}
\title{Auction Results from The Block}
\format{
\subsection{\code{auction_results}}{

A \code{tibble} containing 71 rows and 11 columns:
\describe{
\item{rank}{Auction Rank. 1 indicates winning the auction}
\item{couple}{Couple name, separated by underscore}
\item{reserve}{Reserve price of the property}
\item{auction_result}{Sell price of the property}
\item{profit}{Amount taken home by team. Includes any penalties}
\item{total_winnings}{Total winnings - includes prize money}
\item{auction_order}{Order of the auction. 1 is first}
\item{season}{Show season}
\item{bought_by}{Some notably buyers have their names included. Most are NA}
\item{amount_sold_for_after_auction}{Some properties are sold after auction, their price is put here}
\item{total_profit}{Total amount taken home, includes deductions and prize money. Is often NA as Wikipedia doesn't store it for all seasons.}
}
}
}
\usage{
auction_results
}
\description{
Contains auction results from Seasons 5 to 19 of The Block.
}
\keyword{datasets}
