% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{room_results}
\alias{room_results}
\title{Room Results from The Block}
\format{
\subsection{\code{room_results}}{

A \code{tibble} containing 3236 rows and 6 columns:
\describe{
\item{week}{Week of the show}
\item{area}{Description of the room type}
\item{judge}{Which judge gave the score. Includes "Total" as a judge. Note that
total scores may not be the sum of the individual judge scores, due to penalties or bonus points.}
\item{couple}{Couple name, separated by underscore}
\item{score}{Score for the room, out of 10 (except for Total)}
\item{season}{Show season}
}
}
}
\usage{
room_results
}
\description{
Contains individual room results from Seasons 5 to 19 of The Block
}
\keyword{datasets}
